Move ClickGlob code to new module, explore_tools
Changed from class (as ClickGlob was) to regular module
Now all management of Action and ActionFamily classes is in one place, external to
action module
Encapsulation now much neater
All explore_tools functions are private bar command, command_family and
glob_all which are how the module's accessed (glob_all use comming
shortly)
Now action module is accurately named
How do you feel about explore_tools as a name?
